agent--> dtype

+++ benchmark/paddle/CARLA_SAC/carla_agent.py
@@ -24,13 +24,13 @@ class CarlaAgent(parl.Agent):
         self.alg.sync_target(decay=0)
 
     def predict(self, obs):
-        obs = paddle.to_tensor(obs.reshape(1, -1).astype(np.float32))
+        obs = paddle.to_tensor(obs.reshape(1, -1), dtype='float32')
         action = self.alg.predict(obs)
         action_numpy = action.cpu().numpy()[0]
         return action_numpy
 
     def sample(self, obs):
-        obs = paddle.to_tensor(obs.reshape(1, -1).astype(np.float32))
+        obs = paddle.to_tensor(obs.reshape(1, -1), dtype='float32')
         action, _ = self.alg.sample(obs)
         action_numpy = action.cpu().numpy()[0]
         return action_numpy
@@ -39,9 +39,9 @@ class CarlaAgent(parl.Agent):
         terminal = np.expand_dims(terminal, -1)
         reward = np.expand_dims(reward, -1)
 
-        obs = paddle.to_tensor(obs)
-        action = paddle.to_tensor(action)
-        reward = paddle.to_tensor(reward)
-        next_obs = paddle.to_tensor(next_obs)
-        terminal = paddle.to_tensor(terminal)
+        obs = paddle.to_tensor(obs, dtype='float32')
+        action = paddle.to_tensor(action, dtype='float32')
+        reward = paddle.to_tensor(reward, dtype='float32')
+        next_obs = paddle.to_tensor(next_obs, dtype='float32')
+        terminal = paddle.to_tensor(terminal, dtype='float32')
         self.alg.learn(obs, action, reward, next_obs, terminal)
